requirements.txt
mysql-connector-python==8.0.31
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66
pyzbar==0.1.9
numpy==1.23.4
PyYAML==6.0
confluent-kafka==1.9.2
avro==1.11.1
config-client==1.3.0
requests==2.28.2
